gecko/dom/powersupply/test/test_power_supply_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for PowerSupply API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for PowerSupply API **/

ok('powersupply' in navigator, "navigator.powersupply should exist");

/*
 * TODO: add test cases for powerSupplyOnline/powerSupplyType attributes.
 */
</script>
</pre>
</body>
</html>
